Added seperate flag for incremental analysis + cleanup
