+++ pkg/analysis/dependency_graph.go
@@ -2,7 +2,6 @@ package analysis
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
 	"strings"
 )
@@ -22,50 +21,7 @@ func NewDependencyGraph() *DependencyGraph {
     return dg
 }
 
-func (dg *DependencyGraph) IsInternalDependency(dependency string, rootDir string, language Language) bool {
-    if filepath.IsAbs(dependency) {
-        relPath, err := filepath.Rel(rootDir, dependency)
-        if err != nil {
-            return false
-        }
 
-        return !filepath.IsAbs(relPath) && !strings.HasPrefix(relPath, "..")
-    }
-
-    // Language-specific checks
-    switch language {
-    case LangGo:
-        //extract the moduleName
-        var modulePrefix string
-        goMod := filepath.Join(rootDir, "go.mod")
-        content, err := os.ReadFile(goMod)
-        if err != nil {
-            return false
-        }
-        lines := strings.Split(string(content), "\n")
-        for _, line := range lines {
-            line = strings.TrimSpace(line)
-            if strings.HasPrefix(line, "module "){
-                modulePrefix = strings.TrimSpace(strings.TrimPrefix(line, "module"))
-                dg.moduleName = modulePrefix
-            }else{
-               continue 
-            }
-            
-        }
-
-        return strings.HasPrefix(dependency, modulePrefix)
-    case LangJs, LangTs:
-        // For JS/TS, relative imports are internal
-        return strings.HasPrefix(dependency, "./") || strings.HasPrefix(dependency, "../")
-    case LangPy:
-        // For Python, check if it's a local module
-        // This is a simplified check
-        return !strings.Contains(dependency, ".")
-    }
-    
-    return false
-}
 
 func (dg *DependencyGraph) AddDependency(file, dependency string, rootDir string) {
     file = filepath.Clean(file)
@@ -87,7 +43,7 @@ func (dg *DependencyGraph) AddDependency(file, dependency string, rootDir string
         packageDir := filepath.Join(rootDir, relPath)
         
         // Find all Go files in the package
-        goFiles, err := findGoFilesInPackage(packageDir)
+        goFiles, err := FindGoFilesInPackage(packageDir)
         if err == nil && len(goFiles) > 0 {
             // Add each Go file as a dependency
             for _, goFile := range goFiles {
@@ -131,63 +87,7 @@ func (dg *DependencyGraph) AddDependency(file, dependency string, rootDir string
     dg.dependents[dependency] = append(dg.dependents[dependency], file)
 }
 
-func findGoFilesInPackage(packageDir string) ([]string, error) {
-    var goFiles []string
-    
-    entries, err := os.ReadDir(packageDir)
-    if err != nil {
-        return nil, err
-    }
-    
-    for _, entry := range entries {
-        // Skip directories and non-Go files
-        if entry.IsDir() {
-            continue
-        }
-        
-        name := entry.Name()
-        if !strings.HasSuffix(name, ".go") {
-            continue
-        }
-        
-        // Skip test files
-        if strings.HasSuffix(name, "_test.go") {
-            continue
-        }
-        
-        goFiles = append(goFiles, filepath.Join(packageDir, name))
-    }
-    
-    return goFiles, nil
-}
 
-// GetAffectedFiles returns all files affected by changes in the given files
-func (dg *DependencyGraph) GetAffectedFiles(changedFiles []string) []string {
-    seen := make(map[string]bool)
-    result := make([]string, 0)
-
-    var visit func(string)
-    visit = func(file string) {
-        if seen[file] {
-            return
-        }
-
-        seen[file] = true
-        result = append(result, file)
-
-        // Find all files that depend on this file
-        for _, dependent := range dg.dependents[file] {
-            visit(dependent)
-        }
-    }
-
-    // Start with the changed files
-    for _, file := range changedFiles {
-        visit(file)
-    }
-
-    return result
-}
 
 func (dg *DependencyGraph) GetFileDependencies(path string) ([]string,error) {
     res, ok := dg.dependencies[path]
